Quash compiler warnings by explicitly casting to unsigned on both sides of ? : expressions.
Pointy hat to:	cperciva

SVN-Revision: 3630

+++ libarchive/archive_string_sprintf.c
@@ -64,7 +64,7 @@ append_int(struct archive_string *as, intmax_t d, unsigned base)
 
 	if (d < 0) {
 		archive_strappend_char(as, '-');
-		ud = (d == INTMAX_MIN) ? (uintmax_t)(INTMAX_MAX) + 1 : -d;
+		ud = (d == INTMAX_MIN) ? (uintmax_t)(INTMAX_MAX) + 1 : (uintmax_t)(-d);
 	} else
 		ud = d;
 	append_uint(as, ud, base);
